+++ elements/geosearch/src/main.js
@@ -1,10 +1,47 @@
 import { LitElement, html } from "lit";
+import { unsafeSVG } from "lit/directives/unsafe-svg.js";
 import proj4 from "proj4";
 import _debounce from "lodash.debounce";
 
 import mainStyle from "../../../utils/styles/dist/main.style";
 import buttonStyle from "../../../utils/styles/dist/button.style";
 import { styleEOX } from "./style.eox";
+
+import { getElement } from "../../../utils/getElement";
+
+const loaderSvg = `
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    viewBox="0 0 100 100"
+    preserveAspectRatio="xMidYMid"
+    width="50"
+    height="50"
+    style="shape-rendering: auto; display: block; background: transparent;"
+    xmlns:xlink="http://www.w3.org/1999/xlink"
+  >
+    <g>
+      <circle
+        stroke-dasharray="164.93361431346415 56.97787143782138"
+        r="35"
+        stroke-width="12"
+        stroke="#1a467c"
+        fill="none"
+        cy="50"
+        cx="50"
+      >
+        <animateTransform
+          keyTimes="0;1"
+          values="0 50 50;360 50 50"
+          dur="1.2222222222222223s"
+          repeatCount="indefinite"
+          type="rotate"
+          attributeName="transform"
+        ></animateTransform>
+      </circle>
+      <g></g>
+    </g>
+  </svg>
+`;
 
 class EOxGeoSearch extends LitElement {
   static get properties() {
@@ -30,14 +67,18 @@ class EOxGeoSearch extends LitElement {
        */
       _query: { attribute: false },
       /**
+       * Returns true if the element is currently loading data from the API.
+       * @private
+       */
+      _isLoading: { attribute: false },
+      /**
        * The OpenCage API endpoint to use for the search, including the key but without the query parameter.
        *
        */
       endpoint: { type: String },
-      /**
-       * Selector for the eox-map instance
-       */
-      for: { type: String },
+      for: {
+        type: String,
+      },
       /**
        * The name of the query parameter to use for the search query in the endpoint URI.
        *
@@ -91,6 +132,10 @@ class EOxGeoSearch extends LitElement {
         attribute: "results-direction",
       },
       unstyled: { type: Boolean },
+      loaderSvg: {
+        type: String,
+        attribute: "loader-svg",
+      },
     };
   }
 
@@ -106,17 +151,22 @@ class EOxGeoSearch extends LitElement {
     this._isListVisible = false;
     this._isInputVisible = false;
     this._query = "";
+    this._isLoading = false;
     /**
-     * The query selector for the map
-     * @default eox-map
+     * Query selector of an `eox-map` (`String`, passed as an attribute or property)
+     * or an `eox-map` DOM element (`HTMLElement`, passed as property)
+     *
+     * @type {String|HTMLElement}
      */
     this.for = "eox-map";
     this.listDirection = "right";
     this.resultsDirection = "down";
     this.interval = 800;
+    this.loaderSvg = loaderSvg;
 
     this.fetchDebounced = _debounce(async () => {
-      if (this._query.length <= 2) return;
+      if (this._query.length < 2) return;
+      this._isLoading = true;
       try {
         const uri = `${this.endpoint}${
           this.endpoint.includes("?") ? "&" : "?"
@@ -124,6 +174,7 @@ class EOxGeoSearch extends LitElement {
         const response = await fetch(encodeURI(uri));
         const json = await response.json();
         this._data = json.results;
+        this._isLoading = false;
       } catch (error) {
         console.log("Error setting up or requesting from geosearch endpoint");
       }
@@ -136,6 +187,9 @@ class EOxGeoSearch extends LitElement {
       this._isListVisible = false;
       return;
     } else {
+      if (this._query.length >= 2) {
+        this._isLoading = true;
+      }
       this._isListVisible = true;
     }
     this.fetchDebounced();
@@ -204,7 +258,7 @@ class EOxGeoSearch extends LitElement {
     /**
      * This for now only supports OpenCage
      */
-    const viewProjection = this.#eoxMap.map.getView().getProjection().getCode();
+    const viewProjection = this.eoxMap.map.getView().getProjection().getCode();
 
     let sw = proj4("EPSG:4326", viewProjection, [
       event.bounds.southwest.lng,
@@ -216,7 +270,7 @@ class EOxGeoSearch extends LitElement {
     ]);
     const zoomExtent = [sw[0], sw[1], ne[0], ne[1]];
 
-    this.#eoxMap.zoomExtent = zoomExtent;
+    this.eoxMap.zoomExtent = zoomExtent;
 
     /**
      * The select event, including the details of the selected item
@@ -224,12 +278,39 @@ class EOxGeoSearch extends LitElement {
     this.dispatchEvent(new CustomEvent("geosearchSelect", event));
   }
 
+  updateMap() {
+    const foundElement = getElement(this.for);
+
+    if (foundElement) {
+      const EoxMap = /** @type {import("@eox/map/main").EOxMap} */ (
+        foundElement
+      );
+      this.eoxMap = EoxMap;
+    }
+  }
+
   /**
    * initializes the EOxMap instance
    * And stores it in the private property #eoxMap.
    */
   firstUpdated() {
-    this.#eoxMap = document.querySelector(this.for);
+    this.updateMap();
+  }
+
+  updated(changedProperties) {
+    if (changedProperties.has("for")) {
+      this.updateMap();
+    }
+  }
+
+  get eoxMap() {
+    return this.#eoxMap;
+  }
+
+  set eoxMap(value) {
+    const oldValue = this.#eoxMap;
+    this.#eoxMap = value;
+    this.requestUpdate("eoxMap", oldValue);
   }
 
   render() {
@@ -239,8 +320,16 @@ class EOxGeoSearch extends LitElement {
           display: none;
         }
         ${!this.unstyled && mainStyle}
-        ${!this.unstyled && buttonStyle}
-        ${!this.unstyled && styleEOX}
+          ${!this.unstyled && buttonStyle}
+          ${!this.unstyled && styleEOX}
+          .fill {
+          width: 100%;
+          height: 100%;
+          min-height: 100px;
+          display: flex;
+          justify-content: center;
+          align-items: center;
+        }
       </style>
       <div
         class="geosearch ${this.small ? "small" : ""}"
@@ -289,13 +378,13 @@ class EOxGeoSearch extends LitElement {
             @input="${this.onInput}"
           />
           <ul class="results-container ${this._isListVisible ? "" : "hidden"}">
-            ${this._query.length <= 2
+            ${this._isLoading
+              ? html`<div class="fill">${unsafeSVG(this.loaderSvg)}</div>`
+              : this._query.length < 2
               ? html`<span class="hint"
                   >Enter at least two characters to search</span
                 >`
-              : html``}
-            ${this._query.length >= 2
-              ? this._data.map(
+              : this._data.map(
                   (item) => html`
                     <eox-geosearch-item
                       .item="${item}"
@@ -305,8 +394,7 @@ class EOxGeoSearch extends LitElement {
                       .unstyled=${this.unstyled}
                     />
                   `
-                )
-              : html``}
+                )}
           </ul>
         </div>
       </div>