+++ elements/map/CHANGELOG.md
@@ -1,5 +1,28 @@
 # Changelog
 
+## [1.12.0](https://github.com/EOX-A/EOxElements/compare/map-v1.11.2...map-v1.12.0) (2024-08-06)
+
+
+### Features
+
+* Highlight and animate feature(s) by ID ([#1176](https://github.com/EOX-A/EOxElements/issues/1176)) ([1f95016](https://github.com/EOX-A/EOxElements/commit/1f9501632f00d087473a8deb2435154742f236d2))
+* Loading indicator map control ([#1173](https://github.com/EOX-A/EOxElements/issues/1173)) ([fb727d3](https://github.com/EOX-A/EOxElements/commit/fb727d3fb585216656b9201602fd8de322cc460d))
+* LonLatCenter & lonLatExtent ([#1181](https://github.com/EOX-A/EOxElements/issues/1181)) ([0a154db](https://github.com/EOX-A/EOxElements/commit/0a154db770d262e6ccbb42b786cc2a8d8a5c754f))
+
+## [1.11.2](https://github.com/EOX-A/EOxElements/compare/map-v1.11.1...map-v1.11.2) (2024-07-31)
+
+
+### Bug Fixes
+
+* Issue of layer groups not removing all previous layers ([#1161](https://github.com/EOX-A/EOxElements/issues/1161)) ([e888a61](https://github.com/EOX-A/EOxElements/commit/e888a61c93aafaa73f1130e00b3ab606986407a4))
+
+## [1.11.1](https://github.com/EOX-A/EOxElements/compare/map-v1.11.0...map-v1.11.1) (2024-07-30)
+
+
+### Bug Fixes
+
+* Layer update of group layers ([#1147](https://github.com/EOX-A/EOxElements/issues/1147)) ([5f44bc0](https://github.com/EOX-A/EOxElements/commit/5f44bc0aaaa7b003923f6d51d89ad882bb0fb324))
+
 ## [1.11.0](https://github.com/EOX-A/EOxElements/compare/map-v1.10.1...map-v1.11.0) (2024-07-29)